+++ Lua/NES/Pools.lua
@@ -1,9 +1,3 @@
-
-
-
-
-
-
 --[[
 writeFile: See loadFile for specifics
 --]]
