+++ NES/Cheats/Super Mario Bros.cht
@@ -1,4 +1,2 @@
 075A	02	N	RAM	1		b	h	0	NONE	
-07F8	02	N	RAM	1		b	h	0	NONE	
-07F9	08	N	RAM	1		b	h	0	NONE
